FEDEX/backend/requirements.txt
flask
flask-cors
selenium
pandas
openpyxl
lxml
gunicorn
geopy
webdriver-manager
requests
python-dotenv
